renamed test_dtos dir
