+++ src/test/java/uk/gov/ons/census/bulkprocessor/model/dto/TestPayloadDTO.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.census.bulkprocessor.models.test_dtos;
+package uk.gov.ons.census.bulkprocessor.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 @JsonInclude(Include.NON_NULL)
-public class PayloadDTO {
-  private CollectionCase collectionCase;
-  private RefusalDTO refusal;
+public class TestPayloadDTO {
+  private TestCollectionCase collectionCase;
+  private TestRefusalDTO refusal;
 }
