+++ src/test/java/uk/gov/ons/census/bulkprocessor/model/dto/TestRefusalDTO.java
@@ -1,16 +1,15 @@
-package uk.gov.ons.census.bulkprocessor.models.test_dtos;
+package uk.gov.ons.census.bulkprocessor.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uk.gov.ons.census.bulkprocessor.model.dto.RefusalTypeDTO;
 
 @Data
 @NoArgsConstructor
 @JsonInclude(Include.NON_NULL)
-public class RefusalDTO {
+public class TestRefusalDTO {
   private RefusalTypeDTO type;
-  private CollectionCase collectionCase;
-  private Address address;
+  private TestCollectionCase collectionCase;
+  private TestAddress address;
 }
