+++ src/test/java/uk/gov/ons/census/bulkprocessor/model/dto/TestResponseManagementEvent.java
@@ -1,0 +1,8 @@
+package uk.gov.ons.census.bulkprocessor.model.dto;
+
+import lombok.Data;
+
+@Data
+public class TestResponseManagementEvent {
+  private TestPayloadDTO payload;
+}
