--- src/test/java/uk/gov/ons/census/bulkprocessor/models/test_dtos/ResponseManagementEvent.java
@@ -1,8 +1,0 @@
-package uk.gov.ons.census.bulkprocessor.models.test_dtos;
-
-import lombok.Data;
-
-@Data
-public class ResponseManagementEvent {
-  private PayloadDTO payload;
-}
